requirements.txt
asgiref==3.7.2
black==24.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
flake8==6.1.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
stripe==7.7.0
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
